backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.2
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
pymongo==4.6.0
openpyxl==3.1.2
python-multipart==0.0.6
cryptography==41.0.5
httpx==0.25.2
pytest==7.4.3
alembic==1.12.1 